Allow custom slave connection
This commit allows for custom connection to be passed when
establishing connection to slaves.

While we can assume that with traditional rails apps, a straightforward
host substitution will be suffice, this assumption cannot be made for
projects of a certain scale, where slave boxes may have custom
configurations (maybe different password/username/etc).

This commit's approach relies on the consumer of this package to specify
a lambda that takes in the slave hostname and returns a connection
object.